Prepare to rename the `counter` field to `counter-value`
First we accept either of the field names. We make sure that we define
behaviour when both fields are passed. We could choose to make it an
error, or accept the new field over the old field. In this case, we take
the new field over the old one.
Clients should either be old clients sending the old value, or updated
clients sending the new value.

Using the old name is OK:
```
$ curl --request POST --data '{"counter": 2, "increment-by": 3}' http://localhost:3031/message
{"counter":5}
```

So is using the new name:
```
$ curl --request POST --data '{"counter-value": 2, "increment-by": 3}' http://localhost:3031/message
{"counter":5}
```

Passing both is OK, but we won't encounter clients that do this:
```
$ curl --request POST --data '{"counter-value": 12, "counter": 2, "increment-by": 3}' http://localhost:3031/message
{"counter":15}
```

But we have to pass at least one:
```
$ curl --request POST --data '{"increment-by": 3}' http://localhost:3031/message
Must pass one of "counter" or "counter-value"
```